+++ gulpfile.js
@@ -1,18 +1,21 @@
-var browserSync = require('browser-sync').create();
-var del         = require('del');
-var gulp        = require('gulp');
-var prefixer    = require('gulp-autoprefixer');
-var minifyCss   = require('gulp-clean-css');
-var concat      = require('gulp-concat');
-var eslint      = require('gulp-eslint');
-var imagemin    = require('gulp-imagemin');
-var minifyJson  = require('gulp-jsonminify');
-var sass        = require('gulp-sass');
-var sassLint    = require('gulp-sass-lint');
-var sourcemaps  = require('gulp-sourcemaps');
-var uglify      = require('gulp-uglify');
+const browserSync = require('browser-sync').create();
+const del         = require('del');
+const gulp        = require('gulp');
+const prefixer    = require('gulp-autoprefixer');
+const minifyCss   = require('gulp-clean-css');
+const concat      = require('gulp-concat');
+const eslint      = require('gulp-eslint');
+const gulpIf      = require('gulp-if');
+const minifyImg   = require('gulp-imagemin');
+const minifyJson  = require('gulp-jsonminify');
+const sass        = require('gulp-sass');
+const sassLint    = require('gulp-sass-lint');
+const sourcemaps  = require('gulp-sourcemaps');
+const uglify      = require('gulp-uglify-es').default;
 
-var config      = require('./gulpfile-config.json');
+const config      = require('./gulpfiles/app/gulpfile.json');
+const isEnv       = require('./gulpfiles/app/is-env');
+const lint        = require('./gulpfiles/app/lint');
 
 // processing scss to css and minify result
 function scss() {
@@ -23,21 +26,14 @@ function scss() {
             overrideBrowserslist: ['last 2 versions'],
             cascade: false
         }))
-        .pipe(minifyCss({compatibility: 'ie8'}))
+        .pipe(gulpIf(isEnv(['test', 'prod'], config.env), minifyCss({compatibility: 'ie8'})))
         .pipe(sourcemaps.write('./'))
         .pipe(gulp.dest(config.publicPath + 'css/'));
 }
 
 // lint scss files
 function scssLint() {
-    return gulp.src([
-            config.sourcePath + 'scss/**/*.scss',
-            // exclude third party and special files
-            '!' + config.sourcePath + 'scss/module/_datatables.scss'
-        ])
-        .pipe(sassLint(require('./scss-lint.json')))
-        .pipe(sassLint.format())
-        .pipe(sassLint.failOnError());
+    return lint(gulp, sassLint, [config.sourcePath + 'scss/**/*.scss', '!' + config.sourcePath + 'scss/module/_datatables.scss'], 'scss');
 }
 
 // concatinate and uglify js files
@@ -58,19 +54,14 @@ function js() {
         ])
         .pipe(sourcemaps.init())
         .pipe(concat('scripts.js'))
-        .pipe(uglify())
+        .pipe(gulpIf(isEnv(['test', 'prod'], config.env), uglify()))
         .pipe(sourcemaps.write('./'))
         .pipe(gulp.dest(config.publicPath + 'js/'));
 }
 
 // lint js files
 function jsLint() {
-    return gulp.src([
-            config.sourcePath + 'js/**/*.js'
-        ])
-        .pipe(eslint(require('./js-lint.json')))
-        .pipe(eslint.format())
-        .pipe(eslint.failAfterError());
+    return lint(gulp, eslint, [config.sourcePath + 'js/**/*.js'], 'js');
 }
 
 // copy all json files and minify
@@ -85,11 +76,11 @@ function json() {
 // compress images
 function img() {
     return gulp.src(config.sourcePath + 'img/**/*.{png,gif,jpg,jpeg,ico,xml,json,svg}')
-        .pipe(imagemin([
-            imagemin.gifsicle({interlaced: true}),
-            imagemin.mozjpeg({progressive: true}),
-            imagemin.optipng({optimizationLevel: 5}),
-            imagemin.svgo({
+        .pipe(minifyImg([
+            minifyImg.gifsicle({interlaced: true}),
+            minifyImg.mozjpeg({progressive: true}),
+            minifyImg.optipng({optimizationLevel: 5}),
+            minifyImg.svgo({
                 plugins: [
                     {removeViewBox: true},
                     {cleanupIDs: false}
@@ -115,8 +106,8 @@ function svg() {
 //            'node_modules/@fortawesome/fontawesome-free/sprites/**',
             config.sourcePath + 'svg/**/*.svg'
         ])
-        .pipe(imagemin([
-            imagemin.svgo({
+        .pipe(minifyImg([
+            minifyImg.svgo({
                 plugins: [
                     {removeViewBox: true},
                     {cleanupIDs: false}
@@ -142,6 +133,11 @@ function cleanUp() {
 function browserSyncInit(done) {
     // start browsersync
     browserSync.init({
+        port: 3000,
+        ui: {
+            port: 3001
+        },
+        // ui: false, // enable in production
         proxy: config.localServer
     });
     done();
@@ -171,18 +167,7 @@ function watch() {
 
 // watch files and reload browser on file change
 function watchAndReload() {
-    // watch scss files
-    gulp.watch(config.sourcePath + 'scss/**', gulp.series(scss, scssLint));
-    // watch js files
-    gulp.watch(config.sourcePath + 'js/**', gulp.series(js, jsLint));
-    // watch images
-    gulp.watch(config.sourcePath + 'img/**', img);
-    // watch json files
-    gulp.watch(config.sourcePath + 'json/**', json);
-    // watch fonts
-    gulp.watch(config.sourcePath + 'font/**', font);
-    // watch svg
-    gulp.watch(config.sourcePath + 'svg/**', svg);
+    watch();
     
     gulp.watch(config.publicPath + '**/*.{css,eot,ico,js,json,jpg,otf,png,svg,ttf,woff,woff2}', browserSyncReload);
     gulp.watch('{templates,locale,config,src}/**/*.{php,twig}', browserSyncReload);