+++ gulpfiles/js/module/update-file.js
@@ -9,14 +9,14 @@
 function initUpdateFile() {
     (function($) {
         // 1: active class, 2: label, 3: access icon, 4: value, 5: checked attribute
-        var btnFolderTpl = '<label class="btn btn-primary btn-sm mr-2 mb-2 %s">%s <i class="fas fa-fw fa-check"></i><i class="fas fa-fw fa-times"></i> <i class="fas fa-%s"></i><input value="%s" type="checkbox" class="update-file-folder" %s/></label>';
+        let btnFolderTpl = '<label class="btn btn-primary btn-sm mr-2 mb-2 %s">%s <i class="fas fa-fw fa-check"></i><i class="fas fa-fw fa-times"></i> <i class="fas fa-%s"></i><input value="%s" type="checkbox" class="update-file-folder" %s/></label>';
         
         // 1: button class, 2: active class, 3: label, 4: access icon, 5: value, 6: checked attribute
-        var btnAccessTpl = '<label class="btn btn-%s btn-sm mr-2 mb-2 %s">%s <i class="fas fa-fw fa-check"></i><i class="fas fa-fw fa-times"></i> <i class="fas fa-%s"></i><input value="%s" type="radio" class="update-file-access" name="update_file_access" %s/></label>';
+        let btnAccessTpl = '<label class="btn btn-%s btn-sm mr-2 mb-2 %s">%s <i class="fas fa-fw fa-check"></i><i class="fas fa-fw fa-times"></i> <i class="fas fa-%s"></i><input value="%s" type="radio" class="update-file-access" name="update_file_access" %s/></label>';
         
         $('#update-file').on('show.bs.modal', function(event) {
-            var button = $(event.relatedTarget);
-            var modal = $(this);
+            let button = $(event.relatedTarget);
+            let modal = $(this);
             
             modal.find('.modal-footer button.submit-form').prop('disabled', true);
             modal.find('.btn-create-folder').attr('data-file-id', '');
@@ -56,11 +56,11 @@ function initUpdateFile() {
             
             modal.find('button.submit-form').unbind('click');
             modal.find('button.submit-form').click(function() {
-                var file = {
+                let file = {
                     name: modal.find('#update-file-name').val(),
                     access: modal.find('input.update-file-access:checked').val(),
                     folders: (function() {
-                        var folders = []; // eslint-disable-line array-bracket-newline
+                        let folders = []; // eslint-disable-line array-bracket-newline
                         modal.find('input.update-file-folder:checked').each(function() {
                             folders.push($(this).val());
                         });