+++ src/Entity/FileFolderJoin.php
@@ -13,14 +13,14 @@ class FileFolderJoin extends BaseJoin {
     /**
      * @ORM\Id
      * @ORM\ManyToOne(targetEntity="File", inversedBy="folderJoins")
-     * @ORM\JoinColumn(name="file_id", referencedColumnName="id")
+     * @ORM\JoinColumn(name="file_id", referencedColumnName="id", nullable=false)
      */
     private $file;
     
     /**
      * @ORM\Id
      * @ORM\ManyToOne(targetEntity="Folder", inversedBy="fileJoins")
-     * @ORM\JoinColumn(name="folder_id", referencedColumnName="id")
+     * @ORM\JoinColumn(name="folder_id", referencedColumnName="id", nullable=false)
      */
     private $folder;
     
@@ -91,6 +91,7 @@ class FileFolderJoin extends BaseJoin {
      * Set $file
      * 
      * @param FileFile $file
+     * @return FileFolderJoin
      */
     public function setFile($file) {
         $this->file = $file;
@@ -111,6 +112,7 @@ class FileFolderJoin extends BaseJoin {
      * Set $folder
      * 
      * @param Folder $folder
+     * @return FileFolderJoin
      */
     public function setFolder($folder) {
         $this->folder = $folder;