+++ CHANGELOG.md
@@ -6,8 +6,94 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [3.3.0]
-Please see [`UPGRADE.md`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/UPGRADE.md)!
+## [3.6.0]
+### Added
+- [`gulpfiles/app/lint.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/app/lint.js)
+
+### Changed
+- [`js-lint.json`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/app/js-lint.json)
+- [`scripts.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/js/scripts.js)
+- [`_page.scss`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/scss/module/_page.scss)
+- [`user/show.html.twig`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/templates/user/show.html.twig)
+- [`gulpfile.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfile.js)
+- [`README.md`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/README.md)
+
+### Fixed
+- non spinning fa-sync-alt
+
+## [3.5.0] - 2020-03-09
+Please see [`UPGRADE.md`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/UPGRADE.md#upgrade-from-340-to-350)!
+
+### Added
+- rotated-flipped to [`fontawesome.scss`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/scss/lib/fontawesome.scss)
+- [`_error-animation.scss`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/scss/module/_error-animation.scss)
+- [`version-3.5.0-migration.sql`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/sql/version-3.5.0-migration.sql)
+- [`UploadLimit.php`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Entity/UploadLimit.php) Entity
+- [`UserRepository::getDiskUsage()`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Repository/UserRepository.php)
+- [`partials/error-animation.html.twig`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/templates/partials/error-animation.html.twig)
+- upload limit feature
+- save table state feature
+- reopen tab feature
+- `@return` in PHPDoc for all entities
+- SVG icons from [Fontawesome.com](https://fontawesome.com) for error animation
+
+### Changed
+- [`data-table.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/js/module/data-table.js)
+- [`process-location-hash.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/js/module/process-location-hash.js)
+- [`scripts.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/js/scripts.js)
+- [`styles.scss`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/scss/styles.scss)
+- [`data-model-adminer.png`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/screenshots/data-model-adminer.png)
+- [`data-model.dia`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/screenshots/data-model.dia)
+- [`data-model-dia.png`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/screenshots/data-model-dia.png)
+- [`db-dump.sql`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/sql/db-dump.sql)
+- [`FileController.php`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Controller/FileController.php)
+- [`partials/file-table.html.twig`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/templates/partials/file-table.html.twig)
+- [`partials/folder-table.html.twig`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/templates/partials/folder-table.html.twig)
+- [`user/show-all.html.twig`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/templates/user/show-all.html.twig)
+- [`user/show.html.twig`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/templates/user/show.html.twig)
+- all error pages
+
+### Fixed
+- file extension on download in [`FileController.php`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Controller/FileController.php)
+- application crash on download, if file name has some forbidden special chars
+
+## [3.4.0] - 2020-02-29
+### Added
+- [`upload-file-form.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/js/module/upload-file-form.js)
+- [`data-model-dia.png`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/screenshots/data-model-dia.png)
+- [`data-model.dia`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/screenshots/data-model.dia)
+- [`FileExtension::uniqueFilesQuantity()`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Twig/FileExtension.php)
+- [`gulp-if`](https://www.npmjs.com/package/gulp-if) module
+- `nullable=false` to the most of `@ORM\JoinColumn`
+- ability to save notes without clicking file included button
+
+### Changed
+- [`create-folder.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/js/module/create-folder.js)
+- [`scripts.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/js/scripts.js)
+- [`_timeline.scss`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/scss/module/_timeline.scss)
+- [`npm-postinstall.php`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/php/npm-postinstall.php)
+- [`db-dump.sql`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/sql/db-dump.sql)
+- [`Setup.php`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Composer/Setup.php)
+- [`UserController.php`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Controller/UserController.php)
+- [`.gitignore`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/.gitignore)
+- [`docker-compose.yml`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/docker-compose.yml)
+- [`gulpfile.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfile.js)
+- js files to ES6
+- php translation files
+- margins and paddings of headline in some twig templates
+- vanilla-lazyload 8.17.0 to vanilla-lazyload 12.4.0
+
+### Removed
+- `data-model.png` and replaced by [`data-model-adminer.png`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/screenshots/data-model-adminer.png)
+- gulpfile-config.dist.json and replaced with [`src/app/gulpfile.dist.json`](https://github.com/InsanityMeetsHH/gulp-templating/blob/master/src/app/gulpfile.dist.json)
+- gulp-uglify and replaced with gulp-uglify-es
+
+### Fixed
+- `Undefined index REQUEST_SCHEME` in [`LanguageExtension.php`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Twig/LanguageExtension.php)
+- column Files always "0" in [`user/show-all.html.twig`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/templates/user/show-all.html.twig)
+
+## [3.3.0] - 2020-02-16
+Please see [`UPGRADE.md`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/UPGRADE.md#upgrade-from-320-to-330)!
 
 ### Added
 - [`create-folder.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/js/module/create-folder.js)
@@ -75,8 +161,8 @@ Please see [`UPGRADE.md`](https://github.com/InsanityMeetsHH/file-sharing/blob/m
 - `edit-file.js` and replaced by [`update-file.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/js/module/update-file.js)
 - `modal-edit-file.html.twig` and replaced by [`modal-update-file.html.twig`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/templates/partials/modal-update-file.html.twig)
 
-## [3.2.0]  - 2020-01-19
-Please see [`UPGRADE.md`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/UPGRADE.md)!
+## [3.2.0] - 2020-01-19
+Please see [`UPGRADE.md`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/UPGRADE.md#upgrade-from-310-to-320)!
 
 ### Added
 - 3 access states for files (public, shareable and private).
@@ -99,7 +185,7 @@ Please see [`UPGRADE.md`](https://github.com/InsanityMeetsHH/file-sharing/blob/m
 
 ### Removed
 - `modal.html.twig` and replaced by [`modal-cookie-policy.html.twig`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/templates/partials/modal-cookie-policy.html.twig)
-- `db-model.png` and replaced by [`data-model.png`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/screenshots/data-model.png)
+- `db-model.png` and replaced by [`data-model-dia.png`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/screenshots/data-model-dia.png)
 
 ### Fixed
 - wrong database name in [`docker-compose.yml`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/docker-compose.yml)
@@ -111,7 +197,7 @@ Please see [`UPGRADE.md`](https://github.com/InsanityMeetsHH/file-sharing/blob/m
 - [`GeneralUtility::getUploadMaxFilesize()`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Utility/GeneralUtility.php)
 - `FILE_SHARING_VERSION` in [`public/index.php`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/public/index.php)
 - translations
-- `"type": "project"` in `composer.json`
+- `"type": "project"` in [`composer.json`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/composer.json)
 
 ### Changed
 - jquery 3.3.1 to jquery 3.4.1
@@ -311,7 +397,7 @@ Please see [`UPGRADE.md`](https://github.com/InsanityMeetsHH/file-sharing/blob/m
 ## [2.0.0] - 2018-10-10
 ### Added
 - Upgrade from [Slim Skeleton Fork 4.3](https://github.com/InsanityMeetsHH/Slim-Skeleton/tree/4.3.0) to [Slim Skeleton Fork 5.0.5](https://github.com/InsanityMeetsHH/Slim-Skeleton/tree/5.0.5)
-- Upgrade from [Gulp Templating 3.0](https://github.com/InsanityMeetsHH/gulp-templating/tree/3.0.0) to [Gulp Templating 3.1](https://github.com/InsanityMeetsHH/gulp-templating/tree/3.1.0)
+- Upgrade from [Gulp Templating 3.0](https://github.com/InsanityMeetsHH/file-sharing/tree/3.0.0) to [Gulp Templating 3.1](https://github.com/InsanityMeetsHH/file-sharing/tree/3.1.0)
 - [`cookieconsent.js`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/js/module/cookieconsent.js)
 - [`cookieconsent.scss`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/scss/plugin/cookieconsent.scss)
 - [`_switch.scss`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/gulpfiles/scss/module/_switch.scss)